Move game replay related components to game replay feature folder
