+++ frontend/src/features/gameReplay/components/GameReplayChessboard.tsx
